requirements.txt
Django==1.10.3
diff-match-patch==20121119
django-import-export==0.5.1
requests==2.11.1
tablib==0.11.2
